+++ src/main/java/ru/dsoccer1980/domain/Book.java
@@ -1,19 +1,19 @@
 package ru.dsoccer1980.domain;
 
 public class Book {
-    private int id;
+    private long id;
     private String name;
     private Author author;
     private Genre genre;
 
-    public Book(int id, String name, Author author, Genre genre) {
+    public Book(long id, String name, Author author, Genre genre) {
         this.id = id;
         this.name = name;
         this.author = author;
         this.genre = genre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
