+++ src/main/java/ru/dsoccer1980/shell/GenreCommands.java
@@ -21,7 +21,7 @@ public class GenreCommands {
     @ShellMethod("actions with genres")
     public void genre(
             @ShellOption(defaultValue = "--getAll") String action,
-            @ShellOption(defaultValue = "-1") int id) throws IOException {
+            @ShellOption(defaultValue = "-1") long id) throws IOException {
 
         genreActionService.action(action, id);
